pyproject.toml
[project]
name = "patternrecognition-group1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "matplotlib",
    "scipy",
    "pillow",
    "pandas",
    "torch",
    "torchvision",
    "ipykernel",
    "tqdm",
    "matplotlib",
    "scikit-learn"
]
